Updating api tests
