+++ tests/test_test.py
@@ -1,7 +1,10 @@
-from api.fast import num_remove
+from api import word_transformation
 
 def test_num_remove():
-    assert num_remove('F12') == 'F'
+    assert word_transformation.num_remove('F12') == 'F'
 
-# def test2():
-#     assert False
+def test_punct_remove():
+    assert word_transformation.punct_remove('F.') == 'F'
+
+def test_stop_remove():
+    assert word_transformation.stop_remove('sister and brother') == 'sister brother'
